cleanup and compile-ing on a correctly configured windows system
added some includes of cstdio for printfits hacky and dumb to use
this for C++. strongly consider removing this or decide to code in C

changed makefile to include extra linking targets because mingw
does not by default link to its own libs and there is some
witchcraft going on with the SDL library on windows.

IIRC I did not actually want getError to be a part of anything yet
but, I was bad at git at the time and did not really use branches
how I should have to implement features.

getError.cpp did not compile because there were some wayward chars
removed those.